follow MVC framework that spliting input/output/raise data format to view forder
