+++ model/attraction_model.py
@@ -1,31 +1,9 @@
 from dbconfig import Database
-from model.share import Error
-from pydantic import BaseModel, Field
-from typing import Optional
+from view.share import Error
+from view.attraction_view import Attraction, AttractionListOut, AttractionSingleOut, MrtsOut
 import json
 
 
-# ---------- View ----------
-class Attraction(BaseModel):
-	id: int 
-	name: str 
-	category: str
-	description: str
-	address: str
-	transport: str
-	mrt: str | None = None
-	lat: float
-	lng: float
-	images: list[str]
-class AttractionListOut(BaseModel):
-     nextPage: Optional[int] = Field(None, gt=0)
-     data: list[Attraction]
-class AttractionSingleOut(BaseModel):
-    data: Attraction
-class MrtsOut(BaseModel):
-    data: list
-
-# ---------- Core model  ----------
 class AttractionModel:
 
      def get_attraction_data_list(page: int, keyword: str ="") -> AttractionListOut:
