+++ model/user_model.py
@@ -1,27 +1,18 @@
 from dbconfig import Database
-from model.share import Error
+from view.share import Error
+from view.user_view import Token, CustomizeRaise
+from dotenv import load_dotenv
+
 from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
 from fastapi import Request
-from pydantic import BaseModel, Field
 import bcrypt
 import jwt
 import datetime
 import os
-from dotenv import load_dotenv
 
 load_dotenv()
 
-# ---------- View ----------
-class User(BaseModel):
-    id: int
-    name: str
-    email: str   
-class UserOut(BaseModel):
-    data: User
-class Token(BaseModel):
-    token: str = Field(description="包含JWT加密字串")
-
-# ---------- Core user behavior model  ----------
+# ---------- User behavior model  ----------
 class UserModel:
     def check_email_exist(email: str) -> bool:
         if (Database.read_one("SELECT 1 FROM user WHERE email = %s", (email,))):
@@ -50,16 +41,13 @@ class UserModel:
         token = make_JWT(id=current_user["id"], name=current_user["name"], email=current_user["email"])
         return Token(token=token)
 
+# ---------- JWT setting and model  ----------
+
 SECRET_KEY = os.getenv("SECRET")
 ALGORITHM = "HS256"
 expire = datetime.datetime.now() + datetime.timedelta(days=7)
 
-# ---------- Core JWT validation model  ----------
-class CustomizeRaise(Exception):
-    def __init__(self, status_code, message):
-        self.status_code = status_code
-        self.message = message
-        
+
 class JWTBearer(HTTPBearer):
     def __init__(self, auto_error: bool = False):
         super(JWTBearer, self).__init__(auto_error=auto_error)
@@ -85,7 +73,7 @@ class JWTBearer(HTTPBearer):
         except jwt.InvalidTokenError:
             return {}       
 
-# ---------- something else  ----------
+# --------------------
 
 def make_hash_password(password: str):
     salt = bcrypt.gensalt(rounds=12)
