+++ router/booking.py
@@ -1,33 +1,16 @@
 from fastapi import *
 from fastapi.responses import JSONResponse
-from enum import Enum
-from datetime import  date
-from pydantic import BaseModel
 
-from model.share import Error, Success
-from model.user import JWTBearer
-from model.booking import BookingModel, BookingOut
-
-
+from view.share import Error, Success
+from view.booking_view import BookingInput
+from model.user_model import JWTBearer
+from model.booking_model import BookingModel, BookingOut
 
 router = APIRouter(
      prefix="/api", 
      tags=["Booking"]
 )
-
 security = JWTBearer()
-
-class Time(Enum):
-    morning = "morning"
-    afternoon = "afternoon"
-class Price(Enum):
-    VALUE_2000 = 2000
-    VALUE_2500 = 2500
-class BookingInput(BaseModel):
-    attractionId: int
-    date: date
-    time: Time
-    price: Price
 
 @router.get("/booking", summary="取得尚未確認下單的預定行程", response_model=BookingOut, responses={400:{"model":Error}})
 async def get_booking_data(payload =  Depends(security)):
